feat: save message usage

+++ src/main/java/com/study/studypal/chatbot/entity/MessageUsage.java
@@ -45,5 +45,5 @@ public class MessageUsage {
   private Long outputTokens;
 
   @Column(name = "latency_ms", nullable = false)
-  private Float latencyMs;
+  private Long latencyMs;
 }
